requirements.txt
pyspotify==2.0.0b5
colorama>=0.3.3
mutagen==1.29
requests>=2.3.0